editing pliers interface (i.e., cleaning) to run using iPython
